removed floats in LoRaMessage library #10

+++ lib/LoRaMessage/src/PingMessage.h
@@ -22,8 +22,8 @@ using namespace std;
 class PingMessage : public AbstractMessage{
 	
 	protected:
-		float latitude;
-    float longitude;
+		int latitude;
+        int longitude;
 	
 	public:
     /**
@@ -36,12 +36,12 @@ class PingMessage : public AbstractMessage{
      * @param latitude 
      * @param longitude 
      */
-    PingMessage(const byte recipientId, const byte senderId, const uint32_t messageId, const uint32_t timestamp, const float latitude, const float longitude);
+    PingMessage(const byte recipientId, const byte senderId, const uint32_t messageId, const uint32_t timestamp, const int32_t latitude, const int32_t longitude);
 
     int sendPacket();
 
-    float getLatitude();
-    float getLongitude();
+    int getLatitude();
+    int getLongitude();
 
 };
